graphToken added as global to worker
